Moving connectivity to an independent package
