+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/connectivity/base/BaseConnectivity.java
@@ -16,9 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.traces.connectivity.base;
+package co.elastic.apm.android.sdk.connectivity.base;
 
-import co.elastic.apm.android.sdk.traces.connectivity.Connectivity;
+import co.elastic.apm.android.sdk.connectivity.Connectivity;
 import io.opentelemetry.sdk.metrics.export.MetricExporter;
 import io.opentelemetry.sdk.metrics.export.MetricReader;
 import io.opentelemetry.sdk.trace.SpanProcessor;
